A Good First Program
You should have spent a good amount of time in Exercise 0 learning how
to install a text editor, run the text editor, run the Terminal, and
work with both of them. If you haven't done that, then do not go on. You
will not have a good time. This is the only time I'll start an exercise
with a warning that you should not skip or get ahead of yourself.