Comments and Pound Characters
Comments are very important in your programs. They are used to tell you
what something does in English, and they are used to disable parts of
your program if you need to remove them temporarily.From now on, I'm
going to write code like this. It is important for you to understand
that everything does not have to be literal. Your screen and program may
visually look different, but what's important is the text you type into
the file you're writing in your text editor. In fact, I could work with
any text editor and the results would be the same.